Hack to work around syntax error incompatibilities
Versions of node before 0.12 (notably 0.10) didn't properly propagate
syntax errors.  This also regressed in the node 4.0 releases.

To get around this, if it looks like we are missing the location of the
error, we guess it is (execjs):1

This is obviously not ideal, but only affects syntax errors, and only on
these versions.